fix(#510): lib→components 依存違反を解消し ESLint でガードレール追加
- useRecipes を lib/drip-guide/ から hooks/drip-guide/ に移動
  （React フックなので hooks/ が正しい居場所）
- lib/drip-guide/useRecipes.ts は re-export ブリッジに置き換え
  （@/components/Toast import を除去し依存違反を解消）
- 参照元 4 ページの import パスを @/hooks/drip-guide/useRecipes に更新
- eslint.config.mjs に no-restricted-imports ルールを追加
  （lib/** から @/components/* への import をエラーとして禁止）
- docs/steering/REPOSITORY.md に ESLint 強制の旨を追記

Closes #510
